refactor(relation-dsl): extract filters and selection mixins
Move filter and selection chainers with their normalizers into
relation/dsl/filters.rb and relation/dsl/selection.rb, include them in
DSL, and slim the main DSL while preserving copy-on-write semantics and
behavior

Closes: AE-182
